Feat/bert early stopping (#223)
* change how `clear_cache` is called

* first version of early stopping

* change mypy version

* train_test_split bug fix

* add `compute_metrics` and `EarlyStoppingCallback`

* bug fix

* fix mypy

* try to fix `"eval_f1" not found` error

* forgot to upd `from_context`

* try to fix mypy

* ty to fix "not found f1" error

* refactor a little bit

* disable early stopping for lora

* fix typing errors

* update contributing and makefile

* minor change

* use our metrics

* add docstrings

* set 3.10 for mypy

* upd contributing.md

* try to fix bug

* try to fix typing issue

* try to fix

* add early stopping to ptuning

+++ autointent/modules/scoring/_sklearn/sklearn_scorer.py
@@ -125,9 +125,6 @@ class SklearnScorer(BaseScorer):
         Raises:
             ValueError: If the vector index mismatches the provided utterances
         """
-        if hasattr(self, "_clf"):
-            self.clear_cache()
-
         self._validate_task(labels)
 
         embedder = Embedder(
@@ -165,4 +162,6 @@ class SklearnScorer(BaseScorer):
 
     def clear_cache(self) -> None:
         """Clear cached data in memory used by the embedder."""
-        self._embedder.delete()
+        if hasattr(self, "_clf"):
+            self._embedder.delete()
+            delattr(self, "_clf")
